DS-6373 - Send usage data and determine if cron needs to run.
