+++ modules/custom/social_lets_connect/modules/social_lets_connect_usage/config/install/social_lets_connect_usage.settings.yml
@@ -5,4 +5,4 @@ usage_data:
   system_data: system_data
   modules_installed: modules_installed
 refresh_interval: 86400
-url: ''
+url: 'https://connect.getopensocial.com/usage-data'
