+++ modules/custom/social_lets_connect/modules/social_lets_connect_usage/src/Plugin/ShareUsageDataPluginBase.php
@@ -24,13 +24,6 @@ abstract class ShareUsageDataPluginBase extends PluginBase implements ShareUsage
   public $entityTypeManager;
 
   /**
-   * The date formatter.
-   *
-   * @var \Drupal\Core\Datetime\DateFormatterInterface
-   */
-  public $dateFormatter;
-
-  /**
    * The config.
    *
    * @var \Drupal\Core\Config\ConfigFactory
